deliver require optimizations & fixes
