+++ deliver/lib/deliver/html_generator.rb
@@ -1,3 +1,5 @@
+require_relative 'module'
+
 module Deliver
   class HtmlGenerator
     # Splits keywords supporting:
