+++ deliver/lib/deliver/options.rb
@@ -1,5 +1,8 @@
-require 'fastlane_core'
-require 'credentials_manager'
+require 'fastlane_core/configuration/config_item'
+require 'credentials_manager/appfile_config'
+
+require_relative 'module'
+require_relative 'upload_assets'
 
 module Deliver
   # rubocop:disable Metrics/ClassLength
